+++ src/TypeInferred.HashBang.Html/Extensions.fs
@@ -35,3 +35,5 @@ let (==>) x ys = x |> Element.appendTags ys
 let (-->) x y = x |> Element.appendText [y]
 
 let (+.) x ys = x |> Element.classes ys
+
+let (!!) xs = Div.empty +. xs