+++ src/TypeInferred.HashBang.Owin/ClientSide.fs
@@ -1,14 +1,28 @@
 ﻿module internal TypeInferred.HashBang.Owin.ClientSide
 
 open System
+open Microsoft.FSharp.Quotations
 open FunScript.TypeScript
+open TypeInferred.HashBang
 open TypeInferred.HashBang.Owin.Utilities
 open TypeInferred.HashBang.Routing
 open TypeInferred.HashBang.Html
 
-let createPageScript(handlers, errorPageTemplate) =
+let createPages(pageTypes : System.Type list) =
+    let pageExprs =
+        pageTypes |> List.map (fun pageType ->
+            let consInfo = pageType.GetConstructors().[0]
+            let parameters =
+                consInfo.GetParameters() |> Array.map (fun parameter ->
+                    Expr.Value(null, parameter.ParameterType))
+                |> Array.toList
+            Expr.Coerce(Expr.NewObject(consInfo, parameters), typeof<IPage>))
+    Expr.NewArray(typeof<IPage>, pageExprs)
+
+let createPageScript(pageTypes, errorPageTemplate) =
+    let pages = createPages pageTypes
     <@
-        let handlers = %handlers
+        let pages  = (%%pages : IPage[]) |> Array.toList
         let resources = ref None
 
         let update() =
@@ -18,8 +32,9 @@ let createPageScript(handlers, errorPageTemplate) =
                 if i = -1 then "/"
                 else address.Substring(i + 2, address.Length - i - 2)
             let segments, queryParams = WebUtility.SplitRelativeUri path
+            let request = { Path = segments; QueryParams = queryParams }
             async {
-                let! newHtml = handlers |> List.tryPickAsync (fun f -> f segments queryParams)
+                let! newHtml = pages |> List.tryPickAsync (fun (p : IPage) -> p.TryHandle request)
                 let newHtml =
                     match newHtml with
                     | None -> (%errorPageTemplate) ("Error! Unable to find: " + path)