+++ src/TypeInferred.HashBang.Owin/HashBangMiddleware.fs
@@ -1,11 +1,14 @@
 ﻿namespace TypeInferred.HashBang.Owin
 
 open System
+open System.Reflection
 open System.Threading.Tasks
 open System.Collections.Generic
 open Microsoft.Owin
 open Yahoo.Yui.Compressor
 open FunScript
+open TypeInferred.HashBang
+open TypeInferred.HashBang.SignalR
 open TypeInferred.HashBang.Routing
 open TypeInferred.HashBang.Html
 open TypeInferred.HashBang.MimeTypes
@@ -16,35 +19,49 @@ type Continuation = Func<Environment, Task>
 
 type HashBangMiddleware(next, options:HashBangOptions) =
     inherit OwinMiddleware(next)
-    let funScriptComponents =
-        FunScriptInterop.components @ options.FunScriptComponents
+    let funScriptComponentInjector = fun existingComponents ->
+        options.FunScriptComponentInjector(existingComponents @ FunScriptInterop.components)
 
     let compressor = lazy JavaScriptCompressor()
 
     let compileScript script =
         let isCompressionEnabled = options.IsJavaScriptCompressionEnabled
-        let raw = FunScript.Compiler.Compiler.Compile(script, components = funScriptComponents, noReturn = true, shouldCompress = isCompressionEnabled)
+        let raw = FunScript.Compiler.Compiler.Compile(script, funScriptComponentInjector, noReturn = true, shouldCompress = isCompressionEnabled)
         if isCompressionEnabled then compressor.Value.Compress raw
         else raw
 
+    let pageTypes =
+        options.Pages |> List.map (fun page ->
+            let pageType = page.GetType()
+            match pageType.GetConstructors() with
+            | [|consInfo|] ->
+                let areAllParametersServices =
+                    consInfo.GetParameters() |> Array.forall (fun parameter ->
+                        ServiceEx.isAHashBangService parameter.ParameterType)
+                if not areAllParametersServices then 
+                    failwithf "Pages can only take service types as constructor parameters but found: %s" pageType.FullName
+                else pageType 
+            | _ -> failwithf "Pages must have exactly one constructor."
+        )
+
     let compilePage page =
-        let script = ClientSide.createPageScript(options.PageHandlers, options.ErrorPageTemplate)
+        let script = ClientSide.createPageScript(pageTypes, options.ErrorPageTemplate)
         let compiledScript = compileScript script
         let hashBangScript = Script.empty |> Element.appendText [compiledScript]
-        options.PageTemplate(page, hashBangScript) 
+        options.ServerServedPageTemplate(page, hashBangScript) 
         |> Compiler.compilePage
 
-    let findPage(path, queryParams) =
-        options.CompiledPageHandlers |> List.tryPickAsync (fun handler ->
-            handler path queryParams)
+    let findPage request =
+        options.Pages |> List.tryPickAsync (fun page ->
+            page.TryHandle request)
         |> Async.map (Option.map compilePage)
     
-    let rootPage = findPage([||], Map.empty) |> Async.cache
+    let rootPage = findPage { Path = [||]; QueryParams = Map.empty } |> Async.cache
 
     let getPage url =
         match WebUtility.SplitRelativeUri url with
         | [||], queryParams when queryParams |> Map.isEmpty -> rootPage
-        | path, queryParams -> findPage(path, queryParams)
+        | path, queryParams -> findPage { Path = path; QueryParams = queryParams }
 
     let defaultHeaders = [|
         "Cache-Control", "no-cache"