+++ src/TypeInferred.HashBang.Owin/OwinExtensions.fs
@@ -2,7 +2,9 @@
 
 open System.Runtime.CompilerServices
 open Microsoft.Owin
+open TypeInferred.HashBang
 open TypeInferred.HashBang.Owin
+open TypeInferred.HashBang.SignalR
 
 [<Extension>]
 type AppBuilderExtensions() =
@@ -17,4 +19,10 @@ type AppBuilderExtensions() =
 
     [<Extension>]
     static member UseHashBang<'Html>(appBuilder:IAppBuilder, hashBangOptions:HashBangOptions) =
+        let hashBangOptions = 
+            { hashBangOptions with 
+                FunScriptComponentInjector = fun cs -> 
+                    TypeInferred.HashBang.SignalR.Interop.components 
+                    @ hashBangOptions.FunScriptComponentInjector cs  }
         appBuilder.Use<HashBangMiddleware>(hashBangOptions) |> ignore
+        AppBuilderEx.UseHashBangSignalR(appBuilder, hashBangOptions.Services)