+++ src/TypeInferred.HashBang.SignalR/AppBuilderExtensions.fs
@@ -1,4 +1,4 @@
-﻿namespace Owin
+﻿namespace TypeInferred.HashBang.SignalR
 
 open System
 open System.Reflection
@@ -7,13 +7,11 @@ open Microsoft.Owin
 open Microsoft.AspNet.SignalR
 open TypeInferred.HashBang.SignalR
 
-[<Extension>]
-type AppBuilderExtensions() =
-    [<Extension>]
-    static member UseHashBangSignalR(appBuilder:IAppBuilder, [<ParamArray>]serviceInstances) =
+type AppBuilderEx() =
+    static member UseHashBangSignalR(appBuilder, serviceInstances) =
         GlobalHost.DependencyResolver.Register(typeof<ServiceProtocolConnection>, Func<obj>(fun () -> 
             upcast ServiceProtocolConnection(ServiceCollection(serviceInstances))))
         let route = typeof<ServiceProtocolConnection>.GetCustomAttribute<RouteAttribute>()
-        appBuilder.MapSignalR(route.Path, typeof<ServiceProtocolConnection>, ConnectionConfiguration()) |> ignore
+        Owin.OwinExtensions.MapSignalR(appBuilder, route.Path, typeof<ServiceProtocolConnection>, ConnectionConfiguration()) |> ignore