+++ src/TypeInferred.HashBang.SignalR/Attributes.fs
@@ -1,4 +1,16 @@
-﻿namespace TypeInferred.HashBang.SignalR
+﻿namespace TypeInferred.HashBang
+
+open System
+
+type IService = interface end
+
+module ServiceEx =
+    let private iserviceName = typeof<IService>.FullName
+
+    let isAHashBangService(t : Type) =
+        t.GetInterfaces() |> Seq.exists (fun t -> t.FullName = iserviceName)
+
+namespace TypeInferred.HashBang.SignalR
 
 open System
 
@@ -6,10 +18,6 @@ open System
 type internal RouteAttribute(path : string) =
     inherit Attribute()
     member val Path = path
-
-[<AttributeUsage(AttributeTargets.Class, AllowMultiple=false)>]
-type ServiceAttribute() =
-    inherit Attribute()
 
 [<AttributeUsage(AttributeTargets.Method ||| AttributeTargets.Property, AllowMultiple=false)>]
 type ClientCannotAccessAttribute(path : string) =