+++ src/TypeInferred.HashBang.SignalR/ServiceCollection.fs
@@ -1,7 +1,21 @@
 ﻿namespace TypeInferred.HashBang.SignalR
 
+open System.Reflection
+open TypeInferred.HashBang
+
 /// A collection of services that can be called by name with serialized parameters.
-type internal ServiceCollection(boxedServices) =
+type internal ServiceCollection(boxedServiceInstances : IService list) =
+
+    let boxedServices = 
+        boxedServiceInstances |> Seq.collect (fun obj ->
+            seq {
+                let objType = obj.GetType()
+                yield objType
+                yield! objType.GetInterfaces()
+            }
+            |> Seq.filter ServiceEx.isAHashBangService
+            |> Seq.map (fun t -> t, obj))
+        |> Seq.toArray
 
     let serviceLookup = 
         boxedServices |> Array.map (fun boxedService ->